Consolidate ML queries and harden vector decoding
